requirements.txt
protobuf==3.19.0
numpy==1.23.1
pytorch-lightning=1.8
torch==1.13.0
torchmetrics==0.9.2
torchvision==0.14.0
torchtext==0.14.0
opencv-python==4.5.3.56
Pillow==9.2.0
click==8.0.1
seaborn==0.11.2
